Added ReferenceData, started on TetrisGame
ReferenceData is for arrays of colors and a big array of Tetris pieces.
TetrisGame now has a falling piece that loops to top if it's about to go
out of bounds.

TODO:
Check if a rotation is legal
Switch piece when it hits bottom
Have piece become inert at bottom